# api.yaml
openapi: 3.0.0
info:
  title: Kanban API
  version: 0.2.0
  description: |
    상담 칸반 보드 API. 멀티테넌트 환경 (병원별 로그인 기반).

    - JWT Bearer 인증 (Authorization: Bearer <token>) 사용.
    - `/api/auth/login` 만 public, 그 외 모든 엔드포인트는 인증 필요.
    - 관리자(`is_admin=true`) 계정만 `/api/admin/hospitals` 접근 가능.
    - 날짜별로 컬럼 구조(row1/row2)와 카드를 관리합니다.
    - 템플릿을 통해 컬럼 구조를 다른 날짜에 적용할 수 있습니다.
    - 동일 차트번호 카드 간 형제 동기화(propagation) 기능을 제공합니다.
servers:
  - url: /
    description: Default server
...
security:
  - BearerAuth: []
paths:
...
  /api/admin/hospitals:
    get:
      tags: [admin]
      summary: 병원 목록 조회 (관리자 전용)
      description: |
        등록된 모든 병원의 전체 정보를 조회합니다.
        is_admin=true 계정만 접근 가능. password_hash 는 응답에 절대 포함되지 않습니다.
      responses:
        '200':
          description: 병원 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Hospital'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [admin]
      summary: 병원 신규 등록 (관리자 전용)
      description: |
        새로운 병원(테넌트) 계정을 생성합니다.
        is_admin=true 계정만 접근 가능. 비밀번호는 서버에서 해시되어 저장됩니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HospitalCreate'
            example:
              name: 강남치과
              email: gangnam@example.com
              password: secret1234
      responses:
        '201':
          description: 생성된 병원 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hospital'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/admin/hospitals/{hospital_id}:
    parameters:
      - name: hospital_id
        in: path
        required: true
        schema: { type: string }
        description: 병원 ID
    patch:
      tags: [admin]
      summary: 병원 정보 수정 (관리자 전용)
      description: |
        병원의 이름/비밀번호/활성 상태를 수정합니다.
        is_admin=true 계정만 접근 가능. 비밀번호는 서버에서 해시되어 저장됩니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HospitalUpdate'
            example:
              name: 강남치과(수정)
              is_active: true
      responses:
        '200':
          description: 수정된 병원 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hospital'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [admin]
      summary: 병원 삭제 (관리자 전용 / soft delete)
      description: |
        병원을 비활성화(soft delete) 처리합니다 (is_active=false).
        is_admin=true 계정만 접근 가능.
      responses:
        '204':
          description: 삭제(비활성화) 성공
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/columns/{date}:
    get:
      tags: [columns]
      summary: 일자별 컬럼 구조 조회
      description: 해당 날짜의 row1/row2 컬럼 목록을 반환합니다.
      parameters:
        - $ref: '#/components/parameters/DatePath'
      responses:
        '200':
          description: 컬럼 묶음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ColumnsBundle'
  /api/columns/{date}/row1:
    post:
      tags: [columns]
      summary: row1 컬럼 추가
      description: 해당 날짜에 새로운 row1 컬럼을 추가합니다.
      parameters:
        - $ref: '#/components/parameters/DatePath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ColumnRow1Create'
            example:
              label: 기타상담
              position: 6
      responses:
        '201':
          description: 생성된 row1 컬럼
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ColumnRow1'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/columns/{date}/row1/{r1_id}:
    parameters:
      - $ref: '#/components/parameters/DatePath'
      - name: r1_id
        in: path
        required: true
        schema: { type: string }
        description: row1 컬럼 ID
    patch:
      tags: [columns]
      summary: row1 컬럼 수정
      description: row1 컬럼의 라벨 또는 위치를 수정합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ColumnRow1Update'
            example:
              label: 변경된 라벨
              position: 2
      responses:
        '200':
          description: 수정된 row1 컬럼
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ColumnRow1'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [columns]
      summary: row1 컬럼 삭제
      description: |
        row1 컬럼 및 하위 row2 컬럼, 소속 카드를 삭제합니다.
        built_in=true 컬럼은 삭제 불가합니다.
      responses:
        '204':
          description: 삭제 성공
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/columns/{date}/row2:
    post:
      tags: [columns]
      summary: row2 컬럼 추가
      description: 해당 날짜에 새로운 row2 컬럼을 추가합니다.
      parameters:
        - $ref: '#/components/parameters/DatePath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ColumnRow2Create'
            example:
              row1_id: r1_임플
              label: 추가리프
              position: 3
      responses:
        '201':
          description: 생성된 row2 컬럼
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ColumnRow2'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/columns/{date}/row2/{r2_id}:
    parameters:
      - $ref: '#/components/parameters/DatePath'
      - name: r2_id
        in: path
        required: true
        schema: { type: string }
        description: row2 컬럼 ID
    patch:
      tags: [columns]
      summary: row2 컬럼 수정
      description: row2 컬럼의 라벨, 위치, 소속 row1을 수정합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ColumnRow2Update'
            example:
              label: 변경된 리프
              position: 1
      responses:
        '200':
          description: 수정된 row2 컬럼
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ColumnRow2'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [columns]
      summary: row2 컬럼 삭제
      description: |
        row2 컬럼 및 소속 카드를 삭제합니다.
        built_in=true 컬럼은 삭제 불가합니다.
      responses:
        '204':
          description: 삭제 성공
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/columns/{date}/reorder:
    put:
      tags: [columns]
      summary: 컬럼 재정렬
      description: |
        row1 또는 row2 컬럼의 순서를 일괄 재정렬합니다.
        row1_ids 가 있으면 row1 순서를, row2_ids 가 있으면 각 row1 하위의 row2 순서를 변경합니다.
      parameters:
        - $ref: '#/components/parameters/DatePath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderRequest'
            example:
              row1_ids: [r1_구환, r1_신환, r1_예진, r1_임플, r1_일반, r1_전화]
              row2_ids:
                r1_임플: [r2_임플_상, r2_임플_중, r2_임플_하]
      responses:
        '200':
          description: 재정렬된 컬럼 묶음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ColumnsBundle'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/templates:
    get:
      tags: [templates]
      summary: 템플릿 목록 조회
      responses:
        '200':
          description: 템플릿 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Template'
    post:
      tags: [templates]
      summary: 템플릿 생성
      description: |
        새 템플릿을 생성합니다.
        source_date 가 주어지면 해당 일자의 현재 컬럼 구조를 기반으로 생성합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateCreate'
            example:
              name: 기본 템플릿
              source_date: '2026-04-27'
      responses:
        '201':
          description: 생성된 템플릿
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/templates/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
        description: 템플릿 ID
    get:
      tags: [templates]
      summary: 템플릿 단건 조회
      responses:
        '200':
          description: 템플릿
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [templates]
      summary: 템플릿 수정
      description: 이름이나 row1/row2 구조를 수정합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateUpdate'
            example:
              name: 수정된 템플릿
              row1:
                - { id: r1_구환, label: 구환상담 }
                - { id: r1_신환, label: 신환상담 }
              row2:
                - { id: r2_구환, row1_id: r1_구환, label: 구환 }
                - { id: r2_신환, row1_id: r1_신환, label: 신환 }
      responses:
        '200':
          description: 수정된 템플릿
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [templates]
      summary: 템플릿 삭제
      description: is_default=true 템플릿은 삭제 불가입니다.
      responses:
        '204':
          description: 삭제 성공
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/templates/{id}/duplicate:
    post:
      tags: [templates]
      summary: 템플릿 복제
      description: 기존 템플릿을 복제하여 새 템플릿을 생성합니다.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: 복제 원본 템플릿 ID
      responses:
        '201':
          description: 복제된 템플릿
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        로그인 후 발급받은 JWT 토큰을 `Authorization: Bearer <token>` 헤더에 담아 전송합니다.
  parameters:
    DatePath:
      name: date
      in: path
      required: true
      schema:
        type: string
        format: date
      description: 일자 (YYYY-MM-DD)
      example: '2026-04-27'
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 인증 실패 (토큰 누락/만료/유효하지 않음, 또는 로그인 자격 증명 불일치)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: 권한 부족 (관리자 전용 리소스에 비관리자 계정 접근)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: 제약 충돌 (built_in 삭제, 단일카드 슬롯 충돌, 기본 템플릿 삭제, 이메일 중복 등)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: 에러 코드
          example: NOT_FOUND
        message:
          type: string
          description: 에러 메시지 (사용자 노출용)
          example: 리소스를 찾을 수 없습니다.
        detail:
          type: string
          nullable: true
          description: 상세 디버그 정보 (옵션)
...
    Hospital:
      type: object
      description: |
        병원(테넌트) 전체 정보 (관리자 응답용).
        password_hash 등 민감 필드는 절대 포함하지 않는다.
      required: [id, name, email, is_admin, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          description: 병원 ID
          example: hosp_001
        name:
          type: string
          description: 병원 이름
          example: 강남치과
        email:
          type: string
          format: email
          description: 로그인 이메일
          example: gangnam@example.com
        is_admin:
          type: boolean
          description: 관리자 여부 (true 면 /api/admin/* 접근 가능)
          example: false
        is_active:
          type: boolean
          description: 활성 상태 (false 면 로그인 불가, soft delete 됨)
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    HospitalCreate:
      type: object
      description: 병원 신규 등록 (관리자 전용). 비밀번호는 서버에서 해시되어 저장됨.
      required: [name, email, password]
      properties:
        name:
          type: string
          description: 병원 이름
          example: 강남치과
        email:
          type: string
          format: email
          description: 로그인 이메일 (전체 시스템에서 unique)
          example: gangnam@example.com
        password:
          type: string
          minLength: 8
          description: 초기 비밀번호 (평문, 서버에서 해시 저장)
          example: secret1234
    HospitalUpdate:
      type: object
      description: 병원 정보 부분 수정 (관리자 전용). 모든 필드는 옵션.
      properties:
        name:
          type: string
          nullable: true
          description: 병원 이름
        password:
          type: string
          nullable: true
          minLength: 8
          description: 새 비밀번호 (평문, 서버에서 해시 저장)
        is_active:
          type: boolean
          nullable: true
          description: 활성 상태 토글
...
    ColumnRow1:
      type: object
      required: [id, date, label, position, built_in]
      properties:
        id:
          type: string
          description: row1 컬럼 ID
          example: r1_신환
        date:
          type: string
          format: date
          description: 일자
          example: '2026-04-27'
        label:
          type: string
          description: 컬럼 라벨
          example: 신환상담
        position:
          type: integer
          description: 표시 순서 (0부터 시작)
          example: 1
        built_in:
          type: boolean
          description: 시스템 기본 컬럼 여부 (true 면 삭제 불가)
          example: true
    ColumnRow1Create:
      type: object
      required: [label]
      properties:
        label:
          type: string
          description: 신규 컬럼 라벨
          example: 기타상담
        position:
          type: integer
          nullable: true
          description: 표시 순서 (미지정 시 마지막)
    ColumnRow1Update:
      type: object
      properties:
        label:
          type: string
          nullable: true
        position:
          type: integer
          nullable: true
    ColumnRow2:
      type: object
      required: [id, date, row1_id, label, position, built_in]
      properties:
        id:
          type: string
          example: r2_임플_상
        date:
          type: string
          format: date
        row1_id:
          type: string
          description: 부모 row1 컬럼 ID
          example: r1_임플
        label:
          type: string
          example: 상
        position:
          type: integer
          example: 0
        built_in:
          type: boolean
          example: true
    ColumnRow2Create:
      type: object
      required: [row1_id, label]
      properties:
        row1_id:
          type: string
          description: 부모 row1 컬럼 ID
        label:
          type: string
        position:
          type: integer
          nullable: true
    ColumnRow2Update:
      type: object
      properties:
        label:
          type: string
          nullable: true
        position:
          type: integer
          nullable: true
        row1_id:
          type: string
          nullable: true
          description: 다른 row1 로 이동 시 사용
    ColumnsBundle:
      type: object
      required: [date, row1, row2]
      properties:
        date:
          type: string
          format: date
        row1:
          type: array
          items:
            $ref: '#/components/schemas/ColumnRow1'
        row2:
          type: array
          items:
            $ref: '#/components/schemas/ColumnRow2'
    ReorderRequest:
      type: object
      description: |
        row1 또는 row2 의 정렬 순서를 일괄 지정합니다.
        둘 중 하나 이상이 포함되어야 합니다.
      properties:
        row1_ids:
          type: array
          nullable: true
          description: row1 컬럼 ID의 새 순서
          items:
            type: string
          example: [r1_구환, r1_신환, r1_예진, r1_임플, r1_일반, r1_전화]
        row2_ids:
          type: object
          nullable: true
          description: row1 ID 별 row2 컬럼 ID 의 새 순서
          additionalProperties:
            type: array
            items:
              type: string
          example:
            r1_임플: [r2_임플_상, r2_임플_중, r2_임플_하]
...
    Template:
      type: object
      required: [id, name, row1, row2, is_default, created_at, updated_at]
      properties:
        id:
          type: string
          example: tpl_default
        name:
          type: string
          example: 기본 템플릿
        row1:
          type: array
          items:
            $ref: '#/components/schemas/TemplateColumnItem'
        row2:
          type: array
          items:
            $ref: '#/components/schemas/TemplateLeafItem'
        is_default:
          type: boolean
          description: 시스템 기본 템플릿 여부 (true 면 삭제 불가)
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TemplateColumnItem:
      type: object
      required: [id, label]
      properties:
        id:
          type: string
          example: r1_신환
        label:
          type: string
          example: 신환상담
    TemplateLeafItem:
      type: object
      required: [id, row1_id, label]
      properties:
        id:
          type: string
          example: r2_신환
        row1_id:
          type: string
          example: r1_신환
        label:
          type: string
          example: 신환
    TemplateCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: 새 템플릿
        source_date:
          type: string
          format: date
          nullable: true
          description: 해당 일자의 컬럼 구조를 기반으로 템플릿 생성
    TemplateUpdate:
      type: object
      properties:
        name:
          type: string
          nullable: true
        row1:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/TemplateColumnItem'
        row2:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/TemplateLeafItem'
...